+++ src/Arius.Cli.Tests/DependencyInjectionTests.cs
@@ -36,6 +36,6 @@ public class AddAriusRegistrationTests
         serviceProvider.GetRequiredService<IStreamQueryHandler<ListQuery, RepositoryEntry>>().ShouldBeOfType<ListQueryHandler>();
         serviceProvider.GetRequiredService<IStreamQueryHandler<ChunkHydrationStatusQuery, ChunkHydrationStatusResult>>().ShouldBeOfType<ChunkHydrationStatusQueryHandler>();
         serviceProvider.GetRequiredService<ICommandHandler<ArchiveCommand, ArchiveResult>>().ShouldBeOfType<ArchiveCommandHandler>();
-        serviceProvider.GetRequiredService<ICommandHandler<RestoreCommand, RestoreResult>>().ShouldBeOfType<RestorePipelineHandler>();
+        serviceProvider.GetRequiredService<ICommandHandler<RestoreCommand, RestoreResult>>().ShouldBeOfType<RestoreCommandHandler>();
     }
 }
