+++ src/Arius.Cli.Tests/ProgressTests.cs
@@ -1879,7 +1879,7 @@ public class TrackedDownloadLifecycleTests
         var state = new ProgressState();
 
         // Simulate CreateDownloadProgress adding a TrackedDownload for a large file
-        // Key is RelativePath (the identifier passed from RestorePipelineHandler for large files)
+        // Key is RelativePath (the identifier passed from RestoreCommandHandler for large files)
         var td = new TrackedDownload("photos/sunset.jpg", DownloadKind.LargeFile, "photos/sunset.jpg", compressedSize: 25_400_000, originalSize: 50_000_000);
         state.TrackedDownloads.TryAdd("photos/sunset.jpg", td);
